Flytt dock-skjuling til chat-eid hook med data-attributt
BottomNav lyttet globalt på alle input/textarea-fokuseringer for å skjule
docken — det skapte race-conditions med Portal-fixen og gjorde komponenten
tung å resonnere om. Ny modell: Chat-komponenter eier mekanismen via
useSkjulBottomNavVedFokus(), som setter data-chat-input-fokusert på <html>
når et element med data-chat-input="true" har fokus. CSS skjuler docken.
BottomNav er nå passiv. Portal-fixen fra 636843e er beholdt.
