+++ lib/versjon.json
@@ -1,4 +1,4 @@
 {
-  "nummer": 220,
-  "versjon": "V2.220"
+  "nummer": 221,
+  "versjon": "V2.221"
 }
